# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Risk Management API",
        "description": "RESTful API for Risk Management System",
        "contact": {
            "email": "admin@riskmanagement.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/risks/{risk}/assessments": {
            "get": {
                "tags": [
                    "Risk Assessments"
                ],
                "summary": "Get assessments for a specific risk",
                "operationId": "8a52c2c835ad50dfa364e1e26c4a227d",
                "parameters": [
                    {
                        "name": "risk",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Risk assessments retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Risk Assessments"
                ],
                "summary": "Create a new risk assessment",
                "operationId": "3acbe85db55f8c8eea93ebdb61d1f83c",
                "parameters": [
                    {
                        "name": "risk",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "likelihood_after",
                                    "impact_after",
                                    "assessment_date"
                                ],
                                "properties": {
                                    "likelihood_before": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "impact_before": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "likelihood_after": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "impact_after": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "assessment_notes": {
                                        "type": "string"
                                    },
                                    "assessment_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Risk assessment created successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/risk-assessments/{id}": {
            "get": {
                "tags": [
                    "Risk Assessments"
                ],
                "summary": "Get a specific risk assessment",
                "operationId": "f005d083ba07c43d50e0ac794f4bbb43",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Risk assessment retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Risk Assessments"
                ],
                "summary": "Update a risk assessment",
                "operationId": "5d6131c378c1d6a9efa259e1b01a2f1f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "likelihood_before": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "impact_before": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "likelihood_after": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "impact_after": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "assessment_notes": {
                                        "type": "string"
                                    },
                                    "assessment_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Risk assessment updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Risk Assessments"
                ],
                "summary": "Delete a risk assessment",
                "operationId": "abce81287a0292b7893f39736343b4c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Risk assessment deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}